DEX-951 Multiple response layers in Test API
WIP 3
