+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/docker/DexContainer.scala
@@ -10,10 +10,12 @@ import cats.tagless.implicits._
 import com.dimafeng.testcontainers.GenericContainer
 import com.typesafe.config.Config
 import com.wavesplatform.dex.domain.utils.ScorexLogging
+import com.wavesplatform.dex.it.api.HasWaitReady
+import com.wavesplatform.dex.it.api.responses.dex.MatcherError
 import com.wavesplatform.dex.it.cache.CachedData
 import com.wavesplatform.dex.it.collections.Implicits.ListOps
-import com.wavesplatform.dex.it.dex.{AsyncEnrichedDexNewApi, DexApi, DexNewApi}
-import com.wavesplatform.dex.it.fp
+import com.wavesplatform.dex.it.dex.DexApi.{AsyncTry, SyncTry}
+import com.wavesplatform.dex.it.dex.{AsyncEnrichedDexApi, DexApi}
 import com.wavesplatform.dex.it.resources.getRawContentFromResource
 import com.wavesplatform.dex.it.sttp.LoggingSttpBackend
 import com.wavesplatform.dex.settings.utils.ConfigOps.ConfigOps
@@ -33,12 +35,18 @@ final case class DexContainer private (override val internalIp: String, underlyi
   override protected val cachedRestApiAddress: CachedData[InetSocketAddress] = CachedData(getExternalAddress(DexContainer.restApiPort))
   def restApiAddress: InetSocketAddress = cachedRestApiAddress.get()
 
-  override def api: DexApi[Id] = fp.sync(DexApi[Try](apiKey, restApiAddress))
-  override def asyncApi: DexApi[Future] = DexApi[Future](apiKey, restApiAddress)
+  //override def api: DexApi[Id] = fp.sync(DexApi[Try](apiKey, restApiAddress))
+  //override def asyncApi: DexApi[Future] = DexApi[Future](apiKey, restApiAddress)
 
-  def newAsyncEnrichedApi: AsyncEnrichedDexNewApi = new AsyncEnrichedDexNewApi(apiKey, restApiAddress)
-  def newAsyncApi: DexNewApi[Future] = newAsyncEnrichedApi.mapK(DexNewApi.toAsyncUnsafe)
-  def newApi: DexNewApi[Id] = newAsyncEnrichedApi.mapK(DexNewApi.toSyncUnsafe)
+  def asyncEnrichedApi: AsyncEnrichedDexApi = new AsyncEnrichedDexApi(apiKey, restApiAddress)
+
+  def api: DexApi[Id] = asyncEnrichedApi.mapK(DexApi.toSyncUnsafe)
+  def tryApi: DexApi[SyncTry] = asyncEnrichedApi.mapK(DexApi.toSyncTry)
+  def asyncApi: DexApi[Future] = asyncEnrichedApi.mapK(DexApi.toAsyncUnsafe)
+  def tryAsyncApi: DexApi[AsyncTry] = asyncEnrichedApi.mapK(DexApi.toAsyncTry)
+
+  override def waitReady: HasWaitReady[Id] = ???
+  override def asyncWaitReady: HasWaitReady[Future] = ???
 }
 
 object DexContainer extends ScorexLogging {
