+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/docker/WavesNodeContainer.scala
@@ -9,6 +9,7 @@ import cats.instances.try_._
 import com.dimafeng.testcontainers.GenericContainer
 import com.typesafe.config.Config
 import com.wavesplatform.dex.domain.utils.ScorexLogging
+import com.wavesplatform.dex.it.api.HasWaitReady
 import com.wavesplatform.dex.it.api.node.NodeApi
 import com.wavesplatform.dex.it.cache.CachedData
 import com.wavesplatform.dex.it.fp
@@ -38,8 +39,11 @@ final case class WavesNodeContainer(override val internalIp: String, underlying:
 
   def grpcApiTarget: String = s"${grpcApiAddress.getHostName}:${grpcApiAddress.getPort}"
 
-  override def api: NodeApi[Id] = fp.sync(NodeApi[Try](apiKey, cachedRestApiAddress.get()))
-  override def asyncApi: NodeApi[Future] = NodeApi[Future](apiKey, cachedRestApiAddress.get())
+  def api: NodeApi[Id] = fp.sync(NodeApi[Try](apiKey, cachedRestApiAddress.get()))
+  def asyncApi: NodeApi[Future] = NodeApi[Future](apiKey, cachedRestApiAddress.get())
+
+  override def waitReady: HasWaitReady[Id] = ???
+  override def asyncWaitReady: HasWaitReady[Future] = ???
 
   override def invalidateCaches(): Unit = {
     super.invalidateCaches()
