+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/fp/CanWait.scala
@@ -1,5 +1,6 @@
 package com.wavesplatform.dex.it.fp
 
+import cats.Id
 import com.wavesplatform.dex.it.time.GlobalTimer
 import com.wavesplatform.dex.it.time.GlobalTimer.TimerOpsImplicits
 
@@ -15,9 +16,13 @@ object CanWait {
 
   implicit val future: CanWait[Future] = (duration: FiniteDuration) => GlobalTimer.instance.sleep(duration)
 
-  implicit val tryCanWait: CanWait[Try] = (duration: FiniteDuration) => {
+  implicit val tryCanWait: CanWait[Try] = { (duration: FiniteDuration) =>
     Thread.sleep(duration.toMillis)
     Success(())
   }
+
+  implicit val idCanWait: CanWait[Id] = (duration: FiniteDuration) => Thread.sleep(duration.toMillis)
+
+  def apply[F[_]](implicit W: CanWait[F]): CanWait[F] = W
 
 }
